feat(db): use api key instead of keystore password
Thsi changes whats stored in the users table in the database.

Instead of using the password to the id keystore, this change instead
generates a random 20 character string that it uses as the unique
identifier for each api token for each user.

The locksmith is now changed to give the user that unique id and then
the password to the keystore.

The table for api keys now has the followign information:
- email (why not?)
- the proxy id keystore
- the 20 character unique string for each user

And the api keys now look like this:
<20-char-uniqueid>.<proxy-id-password>

So the proxy id password isn't stored in the database anymore, the only
thing that has it is the user. So when using the API, an API consumer
must know their unique id as well as the password to the keystore
associated with that unique id.

Now if the database is compromised, the hacker will get access to a list
of proxy id keystores but not their passwords.