bugfix(job-deploy): don't consider a job deployed once id is defined...
Originally I left a rase between readystate change and accepted where
if job.id was defined, I'd respond with that jobId. It turns out, although
this isn't technically wrong, if we just wait until the job is accepted,
we'll know for sure that the job is ready to go.

I noticed before this commit, you could query the job with the id when its
in a not ready state. For instance, calling status would return an incorrect
number of total slices..